fix: selectedPage function call, two validaion files, user roles check fixes
