+++ src/main.tsx
@@ -1,4 +1,4 @@
-import { StrictMode } from "react";
+//import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
 import App from "./App.tsx";
@@ -7,9 +7,9 @@ import { Provider } from "react-redux";
 import "./index.css";
 
 createRoot(document.getElementById("root")!).render(
-  <StrictMode>
+  //<StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </StrictMode>
+  //</StrictMode>
 );
